+++ app/components/SupportAssistant/MessageInput.js
