+++ app/pages/page.js
@@ -1,6 +1,6 @@
 'use client'
 
-import { SupportAssistant } from './SupportAssistant';
+import { SupportAssistant } from '../components/SupportAssistant/SupportAssistant';
 
 export default function Home() {
   return (
